add a throwable particle to the first crop item
now I just need to make a way to get these crops from the crop items.
Should be easy enough, have them provide a crop item function, which can
return a crop item if the plant has one. If one is returned while digging,
add it to the players inventory. Just got to make sure this is separate
from the dig items ability